data_loader/requirements.txt
psycopg2==2.9.1
PyYAML==5.4.1